refactor(project structure): Moved files to better structured folders.
